[QMS-342] Read GPX files as POIs
* Added possibility to read GPX files as POIs

+++ src/qmapshack/poi/CPoiPOI.cpp
@@ -355,7 +355,7 @@ void CPoiPOI::slotCheckedStateChanged(QTreeWidgetItem * item)
         return;
     }
 
-    categoryActivated[categoryItem->getId()] = categoryItem->checkState();
+    categoryActivated[categoryItem->getId().toUInt()] = categoryItem->checkState();
 
     loadTimer->start();
 }
